+++ docs/HISTORY.rst
@@ -1,11 +1,15 @@
 Changelog
 =========
 
-0.1.1 (unreleased)
+0.2.0 (unreleased)
 ------------------
 
-- Nothing changed yet.
-
+- Recusively traverse all of the interface's parents attributes
+  [keul]
+- Added an ``IObject`` field adapter. There's no more need that the
+  persistent field implements another interface (see changes in tests).
+  Also: custom IObject derived fields are now automatically supported.
+  [keul]
 
 0.1.0 (2014-05-02)
 ------------------